Moved comments into XML comments, simplified logic in the Custom data attributes, moved attributes to simpler namespace
