+++ src/Stormpath.AspNet/WebApi/AuthorizationFailureResult.cs
@@ -1,13 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
+﻿using System.Net;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using System.Web.Http;
 
-namespace Stormpath.AspNet.Web.Http.Filters
+namespace Stormpath.AspNet.WebApi
 {
     public class AuthorizationFailureResult : IHttpActionResult
     {
