+++ test/Stormpath.AspNet.TckHarness/Controllers/ProfileController.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
-using Stormpath.AspNet.Web.Mvc.Filters;
+using Stormpath.AspNet.Mvc;
 
 namespace Stormpath.AspNet.TckHarness.Controllers
 {
